Remove Twitter image
Unfortunately, showing the logo on Twitter also causes iMessage to pick
it up. It creates a very big link preview that looks a lot less slick
than the default one (which uses the favicon) :-(

Let's leave out the twitter:image for now; hopefully we'll soon have a
better image to use.